perf(storage): document baseline performance before optimization
Record pre-optimization benchmark results for
SET (551.74 RPS, 81.727 ms p50),
GET (745.26 RPS, 61.919 ms p50),
INCR (811.56 RPS, 56.895 ms p50),
LPUSH (688.11 RPS, 66.687 ms p50),
RPUSH (693.68 RPS, 66.175 ms p50).
Provides a reference for comparing performance improvements.
Tested with redis-benchmark on port 6380.

+++ src/main/com/codingchallenges/protocol/Serialise.java
@@ -13,7 +13,7 @@ public class Serialise {
         return "-" + message.trim() + "\r\n"; // Keep it simple, or uppercase if needed
     }
 
-    public static String SerialiseInteger(Integer message) {
+    public static String SerialiseInteger(Long message) {
         if (message == null) {
             return "$-1\r\n"; // Could use nil, but aligning with bulk string for simplicity
         }
@@ -36,7 +36,7 @@ public class Serialise {
         newMessage.append("*").append(message.length).append("\r\n");
         for (Object element : message) {
             if (element instanceof Integer) {
-                newMessage.append(SerialiseInteger((Integer) element));
+                newMessage.append(SerialiseInteger((Long) element));
             } else if (element instanceof String) {
                 newMessage.append(SerialiseBulkString((String) element));
             } else {
